Moved & renamed firebased to root project dir
